+++ example/lib/src/biz/biz1/flutter7/flutter7.page.dart
@@ -14,7 +14,7 @@ class Flutter7Page extends NavigatorStatefulPage {
     super.key,
     required super.moduleContext,
     super.params,
-    super.url,
+    required super.url,
     super.index,
   });