+++ example/lib/src/biz/biz2/flutter2/flutter2.page.dart
@@ -16,7 +16,7 @@ class Flutter2Page extends NavigatorStatefulPage {
     super.key,
     required super.moduleContext,
     super.params,
-    super.url,
+    required super.url,
     super.index,
   });