+++ example/lib/src/biz/biz2/flutter6/flutter6.page.dart
@@ -15,7 +15,7 @@ class Flutter6Page extends NavigatorStatefulPage {
     super.key,
     required super.moduleContext,
     super.params,
-    super.url,
+    required super.url,
     super.index,
   });