+++ lib/src/navigator/navigator_page_lifecycle.dart
@@ -93,7 +93,7 @@ class _NavigatorPageLifecycleState extends State<NavigatorPageLifecycle> {
           _route = route;
 
           _pageObserverCallback = anchor.pageLifecycleObservers.registry(
-            route.settings.url!,
+            route.settings.url,
             _PageLifecyclePageObserver(this),
           );
         }