+++ lib/src/navigator/navigator_page_view.dart
@@ -114,7 +114,7 @@ class _NavigatorPageViewState extends State<NavigatorPageView>
   void _changedToAppear(final RouteSettings routeSettings) {
     if (isAppeared) {
       appearFuture ??= Future.delayed(const Duration(milliseconds: 10), () {
-        final obs = anchor.pageLifecycleObservers[routeSettings.url!];
+        final obs = anchor.pageLifecycleObservers[routeSettings.url];
         for (final ob in obs) {
           if (ob is! _PageViewPageObserver) {
             ob.didAppear(routeSettings);
@@ -128,7 +128,7 @@ class _NavigatorPageViewState extends State<NavigatorPageView>
   void _changedToDisappear(final RouteSettings routeSettings) {
     if (isAppeared) {
       disappearFuture ??= Future.delayed(const Duration(milliseconds: 10), () {
-        final obs = anchor.pageLifecycleObservers[routeSettings.url!];
+        final obs = anchor.pageLifecycleObservers[routeSettings.url];
         for (final ob in obs) {
           if (ob is! _PageViewPageObserver) {
             ob.didDisappear(routeSettings);
@@ -201,7 +201,7 @@ class _NavigatorPageViewState extends State<NavigatorPageView>
         scrollBehavior: widget.scrollBehavior,
         padEnds: widget.padEnds,
         children: widget.routeSettings.map((final it) {
-          var w = ThrioNavigator.build(url: it.url!, params: it.params);
+          var w = ThrioNavigator.build(url: it.url, params: it.params);
           if (w == null) {
             throw ArgumentError.value(
                 it, 'routeSettings', 'invalid routeSettings');