+++ lib/src/navigator/navigator_route_receive_channel.dart
@@ -53,10 +53,10 @@ class NavigatorRouteReceiveChannel {
           'index->${routeSettings.index}',
         );
         routeSettings.params =
-            _deserializeParams(routeSettings.url!, routeSettings.params);
+            _deserializeParams(routeSettings.url, routeSettings.params);
         final animated = arguments?['animated'] == true;
         final callback =
-            anchor.get<NavigatorRoutePushHandle>(url: routeSettings.url!);
+            anchor.get<NavigatorRoutePushHandle>(url: routeSettings.url);
         if (callback != null) {
           final result = await callback(routeSettings, animated: animated);
           if (result == NavigatorRoutePushHandleType.prevention) {